feat: Support version when create / update action (#65)
* feat: Support version when create / update action

The [REST API](https://console.bluemix.net/apidocs/98-ibm-bluemix-openwhisk?=undefined&language=node&env_id=ibm:yp:eu-gb#create-or-update-an-action) support version number, when create or update action. The parameter is optional and if it is empty the system increment the patch version number.

When action invoke the system store the version number in the log. Very useful to know what version of the action have invoked when investigating a bug.

* doc: Add version documentation of create / update action